FIX Don't use actual page instance
Page instances are configured in various ways.
The styleguide controller has it's own Layout template,
and reuses the theme's Page.ss main template.
This can lead to weird situations like showing
content blocks from a random page on top of the
styleguide template, because they happen to be configured on the selected page.

Using Injector::create() is the same logic that's applied in Security controllers
for rendering login forms.